% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_convergence}
\alias{check_convergence}
\title{Check Convergence Criterion}
\usage{
check_convergence(theta_new, theta_old, ll_new, ll_old, tol = 1e-06)
}
\arguments{
\item{theta_new}{New parameter values}

\item{theta_old}{Old parameter values}

\item{ll_new}{New log-likelihood}

\item{ll_old}{Old log-likelihood}

\item{tol}{Parameter convergence tolerance (default: 1e-6)}
}
\value{
Logical, TRUE if converged
}
\description{
Determines if EM algorithm has converged based on parameter and likelihood changes.
}
