% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coef.pdigamma_analysis}
\alias{coef.pdigamma_analysis}
\title{Extract Coefficients}
\usage{
\method{coef}{pdigamma_analysis}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Additional arguments}
}
\description{
Extract Coefficients
}
