% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_initialization.R
\name{compute_dispersion_index}
\alias{compute_dispersion_index}
\title{Sample Dispersion Index Calculation}
\usage{
compute_dispersion_index(y, W = NULL)
}
\arguments{
\item{y}{Vector of counts}

\item{W}{Design matrix for grouping (if NULL, computes global)}
}
\value{
Vector of dispersion indices (sqrt(var/mean))
}
\description{
Computes local dispersion indices for coupling ratio initialization.
}
\details{
Used to initialize rho based on relationship between dispersion and tail thickness
}
