% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fisher_information}
\alias{fisher_information}
\title{Compute Fisher Information Matrix}
\usage{
fisher_information(theta, X = NULL, Z = NULL, W = NULL)
}
\arguments{
\item{theta}{List with alpha, delta, rho (scalars or single values)}

\item{X}{Design matrix for alpha (intercept only if NULL)}

\item{Z}{Design matrix for delta}

\item{W}{Design matrix for rho}
}
\value{
Matrix, Fisher information I(theta)
}
\description{
Calculates expected Fisher information for given parameter values.
Used for standard error estimation and asymptotic theory.
}
\details{
Uses numerical differentiation of log-likelihood
}
\references{
Theorem 2.2, Section 3.2
}
