% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fitted.pdigamma_analysis}
\alias{fitted.pdigamma_analysis}
\title{Extract Fitted Values (Individual Parameters)}
\usage{
\method{fitted}{pdigamma_analysis}(object, parameter = "delta", ...)
}
\arguments{
\item{object}{Fitted model}

\item{parameter}{Which parameter: "alpha", "delta", or "rho"}

\item{...}{Additional arguments}
}
\description{
Extract Fitted Values (Individual Parameters)
}
