% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-Likelihood for PDigamma Regression Model}
\usage{
log_likelihood(y, alpha_vec, delta_vec, rho_vec)
}
\arguments{
\item{y}{Vector of observed counts (positive integers)}

\item{alpha_vec}{Vector of individual alpha parameters}

\item{delta_vec}{Vector of individual delta parameters}

\item{rho_vec}{Vector of individual rho parameters}
}
\value{
Scalar, total log-likelihood
}
\description{
Computes sum of log probabilities for observed data.
}
\details{
All parameter vectors must have length equal to length(y)
}
