% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mean_PDigamma}
\alias{mean_PDigamma}
\title{Mean of PDigamma Distribution}
\usage{
mean_PDigamma(alpha, delta, rho)
}
\arguments{
\item{alpha}{Positive scalar, shape parameter}

\item{delta}{Scalar > 1, total tail index}

\item{rho}{Positive scalar, coupling ratio}
}
\value{
Scalar, E[X] if delta > 2, else Inf
}
\description{
Mean of PDigamma Distribution
}
