% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.pdigamma_analysis}
\alias{predict.pdigamma_analysis}
\title{Predict Method}
\usage{
\method{predict}{pdigamma_analysis}(object, newdata = NULL, type = "parameters", ...)
}
\arguments{
\item{object}{Fitted model}

\item{newdata}{New data frame}

\item{type}{"parameters" (alpha, delta, rho) or "risk" (risk measures)}

\item{...}{Additional arguments}
}
\description{
Predict parameters for new data
}
