% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.pdigamma_analysis}
\alias{print.pdigamma_analysis}
\title{Print Method for PDigamma Analysis}
\usage{
\method{print}{pdigamma_analysis}(x, ...)
}
\arguments{
\item{x}{Object of class pdigamma_analysis}

\item{...}{Additional arguments}
}
\description{
Print Method for PDigamma Analysis
}
