% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_em_core.R
\name{print.pdigamma_fit}
\alias{print.pdigamma_fit}
\title{Print Method for PDigamma Fit}
\usage{
\method{print}{pdigamma_fit}(x, ...)
}
\arguments{
\item{x}{Object of class pdigamma_fit}

\item{...}{Additional arguments}
}
\description{
Print Method for PDigamma Fit
}
