% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qPDigamma}
\alias{qPDigamma}
\title{Quantile Function of PDigamma Distribution}
\usage{
qPDigamma(p, alpha, delta, rho, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities in [0, 1]}

\item{alpha}{Positive scalar, shape parameter}

\item{delta}{Scalar > 1, total tail index}

\item{rho}{Positive scalar, coupling ratio}

\item{lower.tail}{Logical, if TRUE p = P(X <= x), else p = P(X > x) (default: TRUE)}

\item{log.p}{Logical, p is given as log(p) (default: FALSE)}
}
\value{
Vector of quantiles (positive integers)
}
\description{
Computes smallest x such that P(X <= x) >= p.
}
\details{
Uses binary search for efficiency with heavy tails
}
