% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rPDigamma}
\alias{rPDigamma}
\title{Random Number Generation for PDigamma Distribution}
\usage{
rPDigamma(n, alpha, delta, rho)
}
\arguments{
\item{n}{Number of observations}

\item{alpha}{Positive scalar, shape parameter}

\item{delta}{Scalar > 1, total tail index}

\item{rho}{Positive scalar, coupling ratio}
}
\value{
Vector of n random positive integers
}
\description{
Generates random variates using rejection sampling or mixture representation.
}
\details{
Uses latent variable representation (Proposition 2.2):
  U ~ Gamma(alpha, 1), V ~ Gamma(gamma, 1), lambda = U * V^(-1/eta),
  Y|U,V ~ Poisson(lambda)
}
\references{
Proposition 2.2 (Latent Variable Representation)
}
