% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_digamma}
\alias{safe_digamma}
\title{Safe Digamma and Trigamma Functions}
\usage{
safe_digamma(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Vector of digamma values
}
\description{
Wrapper for base::digamma and base::trigamma with input validation.
}
