% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_model_selection.R
\name{store_design_info}
\alias{store_design_info}
\title{Extract Design Matrices from Fit Object}
\usage{
store_design_info(fit, X, Z, W, y)
}
\arguments{
\item{fit}{Fit object}

\item{X}{Original X matrix}

\item{Z}{Original Z matrix}

\item{W}{Original W matrix}

\item{y}{Response vector}
}
\value{
Modified fit object with design_matrices element
}
\description{
Helper to store design matrices for later use (e.g., in diagnostics).
}
