% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.pdigamma_analysis}
\alias{summary.pdigamma_analysis}
\title{Summary Method for PDigamma Analysis}
\usage{
\method{summary}{pdigamma_analysis}(object, ...)
}
\arguments{
\item{object}{Object of class pdigamma_analysis}

\item{...}{Additional arguments}
}
\description{
Summary Method for PDigamma Analysis
}
