% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_em_core.R
\name{summary.pdigamma_fit}
\alias{summary.pdigamma_fit}
\title{Summary Method for PDigamma Fit}
\usage{
\method{summary}{pdigamma_fit}(object, ...)
}
\arguments{
\item{object}{Object of class pdigamma_fit}

\item{...}{Additional arguments}
}
\description{
Summary Method for PDigamma Fit
}
