% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_prep.R
\name{summary_counts}
\alias{summary_counts}
\title{Summary Statistics for Count Data}
\usage{
summary_counts(y)
}
\arguments{
\item{y}{Vector of counts}
}
\value{
List of summary statistics
}
\description{
Computes descriptive statistics useful for model diagnostics.
}
