% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{var_PDigamma}
\alias{var_PDigamma}
\title{Variance of PDigamma Distribution}
\usage{
var_PDigamma(alpha, delta, rho)
}
\arguments{
\item{alpha}{Positive scalar, shape parameter}

\item{delta}{Scalar > 1, total tail index}

\item{rho}{Positive scalar, coupling ratio}
}
\value{
Scalar, Var(X) if delta > 3, else Inf
}
\description{
Variance of PDigamma Distribution
}
